pyproject.toml
[project]
name = "django-fastapi-template"
version = "0.1.0"
description = "A FastAPI x Django template"
authors = [{ name = "kathmandu777", email = "mahhakatomanato@gmail.com" }]
dependencies = [
    "django>=5.1",
    "django-constance>=4.1.1",
    "fastapi>=0.112.2",
    "uvicorn>=0.30.6",
    "pytz>=2024.1",
    "django-debug-toolbar>=4.4.6",
    "django-polymorphic>=3.1.0",
    "setuptools>=74.1.1",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.8.2",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.9",
    "django-model2puml>=0.5.1",
    "fastapi-pagination>=0.12.26",
    "httpx>=0.27.2",
    "requests>=2.32.3",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.mypy]
python_version = "3.11"
show_error_context = true
show_column_numbers = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
plugins = "pydantic.mypy"


[tool.ruff]
target-version = "py311"
line-length = 119
indent-width = 4

[tool.ruff.lint]
# cf. https://docs.astral.sh/ruff/rules/
## ルール
select = ["ALL"] # 全てのチェックを有効にする（Verアップグレードで暗黙的にルールが追加されることは受け入れる）
ignore = [
    "D1",     # D1のみをignoreさせることで、docstringを書いた場合は書いた内容にpydocstyleのルールが適用される
    "TD001",  # invalid-todo-tag: FIXMEも許容するため
    "TD002",  # missing-todo-author: 作者名を書くのが面倒なため
    "TD003",  # missing-todo-link: リンクを貼るのが面倒なため
    "RUF001", # 全角記号などのambiguous unicode characterも使いたいため
    "RUF002", # 全角記号などのambiguous unicode characterも使いたいため
    "RUF003", # 全角記号などのambiguous unicode characterも使いたいため
    "ANN101", # missing-type-self: 引数selfには型ヒントは付けないため
    "ANN102", # missing-type-cls: 引数clsには型ヒントは付けないため
    "G004",   # logging-f-string: loggingでf-stringを使いたいので無視する
    "PGH003", # blanket-type-ignore: 対応が面倒なため
    "PGH004", # blanket-noqa: 対応が面倒なため
    "ANN401", # any-type: anyに名前をつけるのは面倒なため
    "RUF012", # mutable-class-default: Djangoと相性が悪いので無視
    "FBT001",
    "FBT002",
    "S101",   # assertを許可
]

## auto fixの対象
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403", "PLE0604"]
"backend/app/migrations/*.py" = ["E501"]
"backend/app/tests/*.py" = ["S101", "PLR2004"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
